Refactor code to be more clear in intent.
