+++ app/models/concerns/solidus_braintree/add_device_data_concern.rb
@@ -1,0 +1,8 @@
+module SolidusBraintree
+  module AddDeviceDataConcern
+    extend ActiveSupport::Concern
+    included do
+      attr_accessor :device_data
+    end
+  end
+end
