+++ app/models/concerns/solidus_braintree/payment_braintree_nonce_concern.rb
@@ -2,8 +2,7 @@ module SolidusBraintree
   module PaymentBraintreeNonceConcern
     extend ActiveSupport::Concern
     included do
-      attr_accessor :payment_method_nonce
-      attr_accessor :device_data
+      attr_accessor :payment_method_nonce, :device_data
     end
   end
 end
