+++ app/models/concerns/solidus_braintree/permitted_attributes_concern.rb
@@ -1,11 +1,7 @@
 module SolidusBraintree
   module PermittedAttributesConcern
     def payment_attributes
-      super | [:payment_method_nonce, :device_data]
-    end
-
-    def source_attributes
-      super | [:device_data]
+      super | [:payment_method_nonce]
     end
 
     def checkout_attributes
