+++ app/models/order_decorator.rb
@@ -1,0 +1,1 @@
+Spree::Order.include SolidusBraintree::FraudDetectionConcern
