+++ app/models/permitted_attributes_decorator.rb
@@ -1,2 +1,1 @@
 Spree::PermittedAttributes.singleton_class.prepend SolidusBraintree::PermittedAttributesConcern
-Spree::PermittedAttributes.checkout_attributes << :device_data
